Adds fwpl:run annotation for running IO actions.
Currently FWPL is taking no particular precautions. It is the user's
responsibility to not launch the missiles when using FWPL.

Any output generated by the IO action is captured and included in the
rendered VTY view. The result of the action must be showable so that it
can be put on the screen as well.

Currently the IO code is run in the same thread that is compiling and
loading modules. If it hangs, FWPL will hang.